Представления списка избранного текущего пользователя вынесены в отдельный пакет
